Rename daObjKinokoTag_c class header
