+++ include/daObjKinokoTag_c.h
