+++ src/_ZN16daObjKinokoTag_cD0Ev.cpp
@@ -1,7 +1,7 @@
 //cpp
 // @symbol _ZN16daObjKinokoTag_cD0Ev
 
-#include "MegaMushroomCreateTag.h"
+#include "daObjKinokoTag_c.h"
 
 daObjKinokoTag_c::~daObjKinokoTag_c()
 {
