+++ src/_ZN16daObjKinokoTag_cD1Ev.cpp
@@ -1,7 +1,7 @@
 //cpp
 // @symbol _ZN16daObjKinokoTag_cD1Ev
 
-#include "MegaMushroomCreateTag.h"
+#include "daObjKinokoTag_c.h"
 
 daObjKinokoTag_c::~daObjKinokoTag_c()
 {
